+++ core/src/main/scala/japgolly/scalajs/react/vdom/ReactVDom.scala
@@ -172,9 +172,6 @@ object ReactVDom
     override def toString = render.toString
   }
 
-  @deprecated("Nop has been renamed to EmptyTag and will be removed in 0.7.0.", "0.5.0")
-  def Nop = EmptyTag
-
   val EmptyTag: Modifier = new Modifier {
     override def applyTo(t: VDomBuilder): Unit = ()
   }