+++ HISTORY.md
@@ -1,5 +1,6 @@
 ## 0.7.0 (unreleased)
 
+* Removed deprecated methods marked for removal in 0.7.0.
 
 ## 0.6.1 (unreleased)