+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 * Removed deprecated methods marked for removal in 0.7.0.
 * Moved `.experiment` into a new module "extras".
+* `vdom.prefix_<^`
 
 ## 0.6.1 (unreleased)